Defining th
Th is now the collection of theoretical spectra defined in the pickle file. Pay attention in the likelihood: now the offset is added in the likelihood to th which are just the theoretical ones (previously this operation was defined out of the likelihood).

Also added some lines in order to read the pickle fine,  still not sure this is completely correct