Switch to `initialize`
In order to switch to the `initialize` as `Cobaya` recommends, I had to remove any additional param at initialization level.

These must be passed through a `[ClassName].yaml` file, from which it will be read automatically.

This also avoids some tedious behavior when calling the likelihood at the sampler level.