Shift from `Likelihood` to `DatasetLikelihood`
I changed the parent class of `ELICA` from `LIkelihood` to `DatasetLikelihood`.

This allows to automatically get the correct paths to the data.
The "downside" is that data should be in plain text files and not pickled.

Also, the `initialize` method is now handled by the parent class, instead we just have to unpack the info from the ini file.