From Dl to Cl
Changing the spectra to Cl instead of Dl, so it's coherent with Cobaya provider. Also: the fiducial file is just the theoretical power spectrum, and for this reason you have to add the noise_bias and the offset. The Cl data are already biased. Also in this case, the inverse_covmat is provided.