Uploading dataset 100x143
Uploading the files for 100x143 necessary to run the old version of Elica. These files are evaluated starting from the new Cl (June 2025 QML results). The fiducial is the one estimated from the theoretical grid with the cosmology fixed at sroll2 results and with tau=0.06. The offset is evaluated on the new sets of Cls and is calculated using Mangilli's formula. The noise bias is not provided here as it is not used in the old version of the likelihood. The Cl data are unbiased.