Adding the files necessary to run EE_oldWLxWL
These files are created from the simulations obtained in June 2025 (these are not the simulations used in the thesis).
It is provided the inverse covariance matrix. The data are unbiased. The offset is calculated using Mangilli's formula. The noise bias is not provided here, as it was not used in the old version of the thesis.