Updating the log_posterior evaluated with external script for 100x143
This file contains the calculation of the log-posterior using an external script, where the inverse covariance matrix and the offset are computed from the new 250,000 simulations obtained via Cartesian product. In this setup, we test the SH approximation with N_cov = 500. This step is required to validate the likelihood before employing it in the MCMC.