Updating 100x143 files-test no SH approximation
In order to test the likelihood without the SH approximation, the updated dictionary and params.dataset are provided. The only change is in the  "number of simulations", which is set to 0 as a remark that we are not using the SH approximation right now. All the other files are unchanged. The log-posterior evaluated with an external script is also provided.