Updating HybridHL
Updating HybridHL with the covariance matrix evaluated from 250000 simulations (Cartesian product: 500 × 500).
The offset is computed using the quantile method with spectra corresponding to the CMB map and noise map that share the same index.
The file containing the chi2 values, obtained from an external script, is also updated.
This test is performed for the HL case without the SH approximation. For this reason, the number of simulations is set to 0.
When the SH approximation is on, the number of simulations will be set to 1000, matching the expected d.o.f.