Updating hybridHL for paper configuration
The case we want to study in the paper involves the 500 simulations only with the SH approximation with ncov=500. The main update is the inverse covariance matrix, as it is now evaluated from 500 simulations only. The updated logposterior evaluated with an external script is provided for testing the likelihood. The updated dictionary and params.datasets are uploaded.