Updating 100x143 final configuration
Updating the inverse covariance matrix using only 500 simulations (corresponding to the limited number of noise-map realisations available for these channels). The likelihood has been modified to properly handle the SH approximation. Both params.dataset and the associated dictionary are updated in this step. Additionally, a file containing the log-posterior values is provided for testing purposes.